Split up LAT, LLE, GBM catalog classes; added support for reading catalog from disk via fermipy; added support for extended sources
