+++ threeML/catalogs/__init__.py
@@ -1,2 +1,4 @@
-from .Fermi import FermiGBMBurstCatalog, FermiLATSourceCatalog, FermiLLEBurstCatalog
+from .FermiGBM import FermiGBMBurstCatalog
+from .FermiLLE import FermiLLEBurstCatalog
+from .FermiLAT import FermiLATSourceCatalog, FermiPySourceCatalog
 from .Swift import SwiftGRBCatalog
